Redesigned and somewhat functional
After a major disappointment when testing how the database would be
viewed inside the GUI, I've decided to redesign (basically resize) the
frame and everything. Everything is bigger and easier to read.

Also added some functionality: dialog box on closing the application and
populated the database view and notifications view (and more minor
stuff).