Final push
The application is up and ready to go. I have tested it thoroughly and
if any other bugs show up I will make sure to fix them ASAP.

I have also added instructions on how to set everything up as well as a
quick launcher for Windows users and any other missing files (like the
config or message files, which were mentioned in code before but were
never committed).

If there are any problems with the code, feel free to email me.