Review and refactoring
After some initial feedback I have taken another look at the code and
refactored what I considered necessary.

I have taken the liberty of editing the email message and anything
related to it, but this will not be committed (personal use only).
Nevertheless, the code should still work as it is.